requirements.txt
google-genai
discord.py
python-dotenv
better-profanity
